app/src/main/java/com/fahmifuady/jetpack/moviecatalogue/ui/movie/MovieViewModel.kt
package com.fahmifuady.jetpack.moviecatalogue.ui.movie

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.paging.PagedList
import com.fahmifuady.jetpack.moviecatalogue.data.MovieRepository
import com.fahmifuady.jetpack.moviecatalogue.data.source.local.entity.MovieEntity
import com.fahmifuady.jetpack.moviecatalogue.vo.Resource

class MovieViewModel(private val movieRepository: MovieRepository): ViewModel() {
    fun getMovie(): LiveData<Resource<PagedList<MovieEntity>>> = movieRepository.getAllMovie()
}